Register with Email and OTP + JWTs
I added the register endpoint, which sends an email to the user with an OTP that can then be used with the verifyOTP endpoint the set the user to verified in the database. I also went through MERN stack part C, which splits the code into and API.js file and adds Json Web Tokens(I also modified this to work with our recrd database instead of the COP one) for when the user logs in. Also, the password is hashed now and the hashed password works with the login.